# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  POS-fiplex_internal_common.ErrorResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  POS-fiplex_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  POS-fiplex_internal_shift_repository.CashTransactionType:
    enum:
    - cash_in
    - cash_out
    type: string
    x-enum-varnames:
    - CashTransactionTypeCashIn
    - CashTransactionTypeCashOut
  POS-fiplex_internal_shift_repository.ShiftStatus:
    enum:
    - open
    - closed
    type: string
    x-enum-varnames:
    - ShiftStatusOpen
    - ShiftStatusClosed
...
  internal_shift.CashTransactionRequest:
    properties:
      amount:
        minimum: 1
        type: integer
      category:
        type: string
      description:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/POS-fiplex_internal_shift_repository.CashTransactionType'
        enum:
        - cash_in
        - cash_out
    required:
    - amount
    - category
    - type
    type: object
  internal_shift.CashTransactionResponse:
    properties:
      amount:
        type: integer
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      shift_id:
        type: string
      type:
        $ref: '#/definitions/POS-fiplex_internal_shift_repository.CashTransactionType'
      user_id:
        type: string
    type: object
  internal_shift.EndShiftRequest:
    properties:
      actual_cash_end:
        minimum: 0
        type: integer
      password:
        type: string
    required:
    - password
    type: object
  internal_shift.ShiftResponse:
    properties:
      actual_cash_end:
        type: integer
      difference:
        description: Actual - Expected
        type: integer
      end_time:
        type: string
      expected_cash_end:
        type: integer
      id:
        type: string
      start_cash:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/POS-fiplex_internal_shift_repository.ShiftStatus'
      user_id:
        type: string
    type: object
  internal_shift.StartShiftRequest:
    properties:
      password:
        type: string
      start_cash:
        minimum: 0
        type: integer
    required:
    - password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: POS Fiplex API
  title: POS Fiplex API
  version: "1.0"
paths:
...
  /shifts/cash-transaction:
    post:
      consumes:
      - application/json
      description: 'Record a manual cash entry or exit within the active shift (Roles:
        admin, manager, cashier)'
      parameters:
      - description: Cash Transaction Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_shift.CashTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cash transaction created successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_shift.CashTransactionResponse'
              type: object
        "400":
          description: Invalid request body or validation failure
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "404":
          description: No open shift found
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Create a cash transaction (Drop/Expense/In)
      tags:
      - Shifts
      x-roles:
      - admin
      - manager
      - cashier
  /shifts/current:
    get:
      consumes:
      - application/json
      description: 'Check for and retrieve the details of an active shift session
        for the authenticated user (Roles: admin, manager, cashier)'
      produces:
      - application/json
      responses:
        "200":
          description: Open shift retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_shift.ShiftResponse'
              type: object
        "404":
          description: No open shift found
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Get current open shift
      tags:
      - Shifts
      x-roles:
      - admin
      - manager
      - cashier
  /shifts/end:
    post:
      consumes:
      - application/json
      description: 'Close the active shift session for the authenticated user (Roles:
        admin, manager, cashier)'
      parameters:
      - description: End Shift Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_shift.EndShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Shift ended successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_shift.ShiftResponse'
              type: object
        "400":
          description: Invalid request body or validation failure
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "404":
          description: No open shift found
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: End current shift
      tags:
      - Shifts
      x-roles:
      - admin
      - manager
      - cashier
  /shifts/start:
    post:
      consumes:
      - application/json
      description: 'Create a new shift session for the authenticated user (Roles:
        admin, manager, cashier)'
      parameters:
      - description: Start Shift Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_shift.StartShiftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Shift started successfully
          schema:
            allOf:
            - $ref: '#/definitions/POS-fiplex_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_shift.ShiftResponse'
              type: object
        "400":
          description: Invalid request body or validation failure
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "409":
          description: User already has an open shift
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/POS-fiplex_internal_common.ErrorResponse'
      summary: Start a new shift
      tags:
      - Shifts
      x-roles:
      - admin
      - manager
      - cashier
...
swagger: "2.0"